+++ src/PureQL.CSharp.Model/Fields/NullField.cs
@@ -1,0 +1,19 @@
+using PureQL.CSharp.Model.ArrayTypes;
+using PureQL.CSharp.Model.Types;
+
+namespace PureQL.CSharp.Model.Fields;
+
+public sealed record NullField : IField
+{
+    public NullField(string entity, string field)
+    {
+        Entity = entity;
+        Field = field;
+    }
+
+    public string Entity { get; }
+
+    public string Field { get; }
+
+    public IType Type => new NullArrayType();
+}